Move indentation detection to a separate class
